+++ src/main/java/com/example/team1_be/domain/User/Role/Service/RoleService.java
@@ -1,4 +1,4 @@
-package com.example.team1_be.domain.User.Role;
+package com.example.team1_be.domain.User.Role.Service;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
