+++ src/main/java/com/example/team1_be/domain/User/Role/Service/RoleWriteOnlyService.java
@@ -1,14 +1,16 @@
-package com.example.team1_be.domain.User.Role;
+package com.example.team1_be.domain.User.Role.Service;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+
+import com.example.team1_be.domain.User.Role.RoleRepository;
 
 import lombok.RequiredArgsConstructor;
 
 @Service
 @RequiredArgsConstructor
-@Transactional(readOnly = true)
-public class RoleReadOnlyService {
+@Transactional
+public class RoleWriteOnlyService {
 	private final RoleRepository repository;
-	
+
 }
